Split collections and fixed helper's init
