+++ praw/models/helpers/__init__.py
@@ -1,5 +1,5 @@
 """Submodule for all helpers."""
-
-from .live import LiveHelper  # noqa: F401
-from .multireddit import MultiredditHelper  # noqa: F401
-from .subreddit import SubredditHelper  # noqa: F401
+# flake8: noqa
+from .live import LiveHelper
+from .multireddit import MultiredditHelper
+from .subreddit import SubredditHelper
